gnn-backend/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.18.2
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
pycountry==24.6.1
